Remove node, yarn, gulp, vue, etc from Graphene
All of these items are currently unused as part of the build process, and are being removed.  (Many of these dependencies flag security errors in git, which is not ideal when considering that they are unused parts of the project)

If we want to add this stuff back in at a later date, we can pull the necessary files directly from a clean laravel template.